+++ app/Http/Controllers/HomeController.php
@@ -16,7 +16,7 @@ class HomeController extends Controller
     public function __construct()
     {
         parent::__construct();
-        $this->middleware('auth');
+        // $this->middleware('auth');
     }
 
     public function getGlobalData() {