+++ routes/web.php
@@ -11,6 +11,6 @@
 |
 */
 
-Route::get('/', 'HomeController@home')->name('home');
+Route::get('/', 'HomeController@index')->name('home');
 
 Auth::routes();