+++ webpack.mix.js
@@ -53,3 +53,6 @@ mix.js('resources/js/app.js', 'public/js')
    });
    // TODO wait for webpack 5, since css extraction does not work with v4 and extract()
    // .extract();
+if(appPath) {
+    mix.copyDirectory('public/' + appPath + 'fonts', 'public/fonts');
+}