+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.6.1
+### Added
+- Drag & Drop
+- Info modal when importing a new file
+### Fixed
+- Skip resources that are not concepts of imported files
+- Wrong URLs (Add/Delete broader)
+- Font loading
+- Remove login page redirect on page load
+- Prevent adding broaders as narrowers (would create circles)
+- Several bugs in add/remove concepts
+- Minor style fixes
+- Logout after typing too fast in tree/concept search
+
 ## 0.6 - Federsee
 This version is a complete rewrite using Laravel and Vue.js. Please refer to the [INSTALL.md](INSTALL.md) for migration and new setup information.